Modify the turtle to use matrices to compute pos
Switch from computing the turtle position "manually" to using Matrix4x4 to
deal with the movement. The code is much simplified because of this.

To be noted, many resulting example SVG files have been modified because of
how the rounding is done. It used to be that the position was rounded after
every movement. Now only the values that are sent to be rendered are rounded;
the internal position is never rounded. The results are practically the same
and the differences are tiny (of the order of about 0.0001).

+++ src/RubberPlant/LSystem/LSystem.cs
@@ -8,7 +8,7 @@ namespace RubberPlant
     public class LSystem
     {
         public string Name { get; set; }
-        public double Angle { get; set; }
+        public float Angle { get; set; }
         public Rule Axiom { get; set; } = new Rule();
         public List<Rule> Rules { get; set; } = new List<Rule>();
         public Dictionary<Atom, TurtleCommand> Vocabulary { get; set; }
